Убраны лишние зависимости
