+++ app/src/main/java/net/svishch/android/githubclient/di/modules/DatabaseModule.kt
@@ -4,9 +4,12 @@ import androidx.room.Room
 import dagger.Module
 import dagger.Provides
 import net.svishch.android.githubclient.App
+import net.svishch.android.githubclient.mvp.model.ModelData
+import net.svishch.android.githubclient.mvp.model.ModelDataProviders
 import net.svishch.android.githubclient.mvp.model.entity.room.Database
 import net.svishch.android.githubclient.mvp.model.cache.IGithubRepositoriesCache
 import net.svishch.android.githubclient.mvp.model.cache.IGithubUsersCache
+import net.svishch.android.githubclient.mvp.model.db.DataDb
 import ru.geekbrains.githubclient.mvp.model.cache.room.RoomGithubRepositoriesCache
 import ru.geekbrains.githubclient.mvp.model.cache.room.RoomGithubUsersCache
 
@@ -28,5 +31,11 @@ class DatabaseModule {
     @Provides
     fun repositoriesCache(database: Database): IGithubRepositoriesCache {
         return RoomGithubRepositoriesCache(database)
+    }
+
+    @Singleton
+    @Provides
+    fun dataDb(database: Database): DataDb {
+        return DataDb(database)
     }
 }
