+++ app/src/main/java/net/svishch/android/githubclient/mvp/presenter/InfoPresenter.kt
@@ -3,8 +3,13 @@ package net.svishch.android.githubclient.mvp.presenter
 import moxy.MvpPresenter
 import net.svishch.android.githubclient.mvp.view.InfoView
 import ru.terrakok.cicerone.Router
+import javax.inject.Inject
 
-class InfoPresenter(val router: Router) : MvpPresenter<InfoView>() {
+class InfoPresenter() : MvpPresenter<InfoView>() {
+
+    @Inject
+    lateinit var router: Router
+
     override fun onFirstViewAttach() {
         super.onFirstViewAttach()
         viewState.init()
